pyproject.toml
[project]
name = "PROJECT_NAME"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fake-bpy-module>=20251003",
]

[dependency-groups]
dev = [
    "python-lsp-server>=1.13.1",
    "ruff>=0.14.0",
]
